pyproject.toml
[tool.poetry]
name = "valuation-project"
version = "0.1.0"
description = "valuing mutually exclusive capital projects"
authors = ["hy"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.1.1"
streamlit = "^1.27.2"
streamlit-lottie = "^0.0.5"
ipykernel = "^6.25.2"
st-pages = "^0.4.5"
scipy = "^1.11.3"
streamlit-aggrid = "^0.3.4.post3"
st-annotated-text = "^4.0.1"
streamlit-extras = "^0.3.4"
lightgbm = "^4.1.0"
plotly = "^5.17.0"
streamlit-pills = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
